Update README files for examples
I wasn't able to run the sessionize script against pig 0.12.1.  It appears there is a bug in ToString that isn't fixed until 0.13.0.  Sessionize accepts ms time as the input in addition to iso time anyways though, so I removed this code that converted to an iso string.

I also updated the quantile's data generation script to output a file with a tsv extension.