DATAFU-76 Resolve issues with build-plugin directory
https://issues.apache.org/jira/browse/DATAFU-76

I've attempted to clean up the NOTICE and LICENSE files.  For reference I read http://www.apache.org/dev/licensing-howto.html.  I removed everything form the NOTICE because I don't think they were actually required to be there.  Many are not actually bundled, and for the ones that are they do not have a NOTICE, so there is nothing that needs to be propagated in our NOTICE file.  The one exception is commons-math, which does have a NOTICE file.  I still need to update the NOTICE with those files that we do include from there.

I also updated the LICENSE file to include licensing info for files that we bundle.