Update documentation and fix various issues
I've reviewed the website content and made many changes to make sure it reflects the current state of the code base.  Most of the changes have to do with updating the Hourglass instructions.  I also made some changes to the Hourglass build so support the demo, which was lost in the Gradle migration away from Ant.  I've also fixed many of the links so they point to the right location.