DATAFU-95 Improve wrong JDK error message
Use @SupportedSourceVersion(SourceVersion.RELEASE_7) in MultilineProcessor.java.  It doesn't seem there was any need to require RELEASE_8.
Ensure build-plugin is built (except when releasing to maven), as otherwise we can't run tests in Eclipse.
Add sweeney to check JDK 1.7 is used.
Bump the datafu version to 1.3.1 as this was missed after the last release.