DATAFU-146 Upgrade to Gradle v4.8.1
This updates to Gradle 4.8.1.  The only issue I encountered doing this was that autojar 1.0.1 is not compatible with this version of Gradle.  Only a small modification is required to fix it:

    manifest.writeTo(manifestFile.path)

The autojar plugin does not appear to be maintained anymore, so I have checked the code into our repo.  It is Apache 2 licensed so this should not be a problem.  Since we can't bundle the autojar-2.1.jar I have kept the dependency on gradle-autojar-1.0.1, as this has it bundled as a resource.  I modified ExtractAutojar in the plugin to account for this difference.

Signed-off-by: Eyal Allweil <eyal@apache.org>