DATAFU-136 Update release instructions for proper handling of release flag
The new approach is to create an x.y.z-rc0 branch from x.y.z and set release=true in the release candidate branch.  This way when you checkout the branch/tag for the release candidate you can build with the same instructions as when you use the source release.